feat(codecov): Update url when changing option radio
