+++ static/app/routes.tsx
@@ -2022,9 +2022,7 @@ function buildRoutes() {
         {/* Render tests onboarding without any layout wrapping */}
         <Route
           path="new/"
-          component={make(
-            () => import('sentry/views/codecov/tests/onboarding/onboarding')
-          )}
+          component={make(() => import('sentry/views/codecov/tests/onboarding'))}
         />
       </Route>
       <Route path="tokens/">
