+++ static/app/views/codecov/tests/onboarding.tsx
@@ -1,13 +1,46 @@
+import {useCallback} from 'react';
+import {useSearchParams} from 'react-router-dom';
 import styled from '@emotion/styled';
 
+import RadioGroup from 'sentry/components/forms/controls/radioGroup';
+import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import TestPreOnboardingPage from 'sentry/views/codecov/tests/preOnboarding';
+
+type SetupOption = 'githubAction' | 'cli';
 
 export default function TestsOnboardingPage() {
+  const [searchParams, setSearchParams] = useSearchParams();
+  const opt = searchParams.get('opt');
+  const handleRadioChange = useCallback(
+    (newOption: SetupOption) => {
+      setSearchParams({opt: newOption});
+    },
+    [setSearchParams]
+  );
+
   return (
     <LayoutGap>
       <p>Test Analytics Onboarding</p>
-      <TestPreOnboardingPage />
+      <OnboardingContainer>
+        <IntroContainer>
+          <GetStartedHeader>{t('Get Started with Test Analytics')}</GetStartedHeader>
+          <TAValueText>
+            {t(
+              'Test Analytics offers data on test run times, failure rates, and identifies flaky tests to help decrease the risk of deployment failures and make it easier to ship new features quickly.'
+            )}
+          </TAValueText>
+        </IntroContainer>
+        <SelectOptionHeader>Select a setup option</SelectOptionHeader>
+        <RadioGroup
+          label="Select a setup option"
+          value={opt === 'cli' ? 'cli' : 'githubAction'}
+          onChange={handleRadioChange}
+          choices={[
+            ['githubAction', t('Use GitHub Actions to run my CI')],
+            ['cli', t("Use Sentry Prevent's CLI to upload testing reports")],
+          ]}
+        />
+      </OnboardingContainer>
     </LayoutGap>
   );
 }
@@ -15,4 +48,32 @@ export default function TestsOnboardingPage() {
 const LayoutGap = styled('div')`
   display: grid;
   gap: ${space(2)};
+`;
+
+const OnboardingContainer = styled('div')`
+  padding: ${space(1.5)} ${space(4)};
+  border: 1px solid ${p => p.theme.border};
+  border-radius: ${p => p.theme.borderRadius};
+`;
+
+const IntroContainer = styled('div')`
+  border-bottom: 1px solid ${p => p.theme.border};
+  padding-bottom: ${space(3)};
+`;
+
+const GetStartedHeader = styled('h2')`
+  font-size: 1.625rem;
+  color: ${p => p.theme.tokens.content.primary};
+  line-height: 40px;
+`;
+
+const TAValueText = styled('p')`
+  font-size: ${p => p.theme.fontSizeLarge};
+  color: ${p => p.theme.tokens.content.primary};
+`;
+
+const SelectOptionHeader = styled('h5')`
+  font-size: ${p => p.theme.fontSizeExtraLarge};
+  color: ${p => p.theme.tokens.content.primary};
+  margin-top: ${space(3)};
 `;
