+++ static/app/views/codecov/tests/tests.tsx
@@ -4,7 +4,7 @@ import {DatePicker} from 'sentry/components/codecov/datePicker/datePicker';
 import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import PageFiltersContainer from 'sentry/components/organizations/pageFilters/container';
 import {space} from 'sentry/styles/space';
-import TestsOnboardingPage from 'sentry/views/codecov/tests/onboarding/onboarding';
+import TestsOnboardingPage from 'sentry/views/codecov/tests/onboarding';
 import {Summaries} from 'sentry/views/codecov/tests/summaries/summaries';
 
 const DEFAULT_CODECOV_DATETIME_SELECTION = {
